Refactor: Move Providers from main to App
